ai-engine/requirements.txt
flask==3.0.2
flask-cors==4.0.0
stockfish==3.28.0
pydantic==2.5.3
requests>=2.32.3,<3.0.0
openai==1.6.1
python-dotenv==1.0.0
httpx==0.24.1 